removing more warnings

+++ src/stan/prob/distributions/multivariate/continuous/lkj_corr.hpp
@@ -24,7 +24,7 @@ namespace stan {
       if (eta == 1.0) {
         // C++ integer division is appropriate in this block
         Eigen::VectorXd numerator( Km1 / 2 );
-        for(size_t k = 1; k <= numerator.rows(); k++)
+        for(int k = 1; k <= numerator.rows(); k++)
           numerator(k-1) = lgamma(2 * k);
         constant = sum(numerator);
         if ( (K % 2) == 1 ) constant += 0.25 * (K * K - 1) * LOG_PI -
@@ -35,7 +35,7 @@ namespace stan {
       }
       else {
         constant = -Km1 * lgamma(eta + 0.5 * Km1);
-        for (size_t k = 1; k <= Km1; k++)
+        for (int k = 1; k <= Km1; k++)
           constant += 0.5 * k * LOG_PI + lgamma(eta + 0.5 * (Km1 - k));
       }
       return constant;
